Fixes Input and Output
Removes output for header bringing the 'find_avg' variable (which counts how many home team wins) down to 152 (which should now be the correct value) and brings the 'total_counter' variable (which counts how many games have been played) down to 267 (the correct value) and enables the removal of the value of '-1' as the starting value of 'total_counter' and instead replaces it with 0. Essentially this just improves the accuracy. The commit makes a -0.4% change in the final output.